feat: implement luck index and release 0.4.8
We introduce the Luck Index, an estimate based on the sequence of piece reveals that measures the degree of luck in revealing pieces, regardless of their positions on the board. A positive value indicates good luck for red, while a negative value indicates good luck for black. This estimation model is derived from 8,345 VSTC test games.

Note that red inherently has a tempo advantage at the start, so the Luck Index defaults to a positive value, as obtaining the first move is itself a form of good luck.